feat: implement sensor and clock classes with header files for DustSensor, FrontSensor, LeftSensor, RightSensor, and DigitalClock
