+++ 2026_1_OOAD/HDWARE/LeftSensor.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "Class/ISensor.h"
+
+class LeftSensor : public ISensor {
+private:
+    float detectedValue = -1.0F;
+
+public:
+    bool detect() override {
+        return detectedValue > 0 && detectedValue < 5;
+    }
+
+    void setLeftSensorValue(float value) {
+        detectedValue = value;
+    }
+};
