--- 2026_1_OOAD/HDWARE/Motor.cpp
@@ -1,22 +1,0 @@
-//
-//
-//  Generated by StarUML(tm) C++ Add-In
-//
-//  @ Project : OOAD5T
-//  @ File Name : Motor.cpp
-//  @ Date : 2026-05-03
-//  @ Author :
-//
-//
-
-#include "Motor.h"
-
-void Motor::stop() {}
-
-void Motor::moveForward() {}
-
-void Motor::turnLeft() {}
-
-void Motor::turnRight() {}
-
-void Motor::moveBackward() {}
